+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-1.7.0
------------------
+1.7.0 (2017-10-30)
+------------------
 
 * The following payment processors were added to the `Payment.Processor` enum:
   * `BPOINT`